+++ apps/spa-3ncount3r/src/app/App.tsx
@@ -5,7 +5,7 @@ import {
 } from "react-router-dom";
 import { CssBaseline, ThemeProvider } from '@mui/material';
 
-import router from './Router';
+import router from '../Router';
 import { theme } from './theme'
 
 class App extends React.Component {
